Print timestamp

+++ src/train/training_pipeline.py
@@ -1,23 +1,36 @@
 import subprocess
 import os
 import shutil
+import time
+import datetime
 
 if __name__=="__main__":
     # params
     parent_dir = os.path.dirname(os.path.abspath(__file__))
 
+    # clean up bag2labels/output_data
+    if os.path.exists(parent_dir+"/bag2labels/output_data"):
+        print("Deleting folder: {}".format(parent_dir+"/bag2labels/output_data"))
+        shutil.rmtree(parent_dir+"/bag2labels/output_data")
+
     # run bag2labels.py
     print("Running bag2labels/bag2labels.py")
-    subprocess.run(["python3",parent_dir+"/bag2labels/bag2labels.py"])
+    result = subprocess.run(["python3",parent_dir+"/bag2labels/bag2labels.py"], stdout=subprocess.PIPE)
+    print("Running bag2labels.py complete at {}".format(datetime.datetime.now()))
 
     # clean up train/input_data
-    print("Deleting folder: {}".format(parent_dir+"/train/input_data"))
-    shutil.rmtree(parent_dir+"/train/input_data")
+    if os.path.exists(parent_dir+"/train/input_data"):
+        print("Deleting folder: {}".format(parent_dir+"/train/input_data"))
+        shutil.rmtree(parent_dir+"/train/input_data")
 
     # copy files from bag2labels/output_data to train/input_data
-    print("Copying files form {} to {}".format(parent_dir+"/bag2labels/output_data", parent_dir+"/train/input_data"))
+    print("Copying files from {} to {}".format(parent_dir+"/bag2labels/output_data", parent_dir+"/train/input_data"))
     shutil.copytree(parent_dir+"/bag2labels/output_data", parent_dir+"/train/input_data")
+    print("Copying files complete at {}".format(datetime.datetime.now()))
 
     # run train.py
-    print("Running train/train.py")
-    subprocess.run(["python3",parent_dir+"/train/train.py"])
+    if result.returncode==0:
+        print("Running train/train.py {}".format(datetime.datetime.now()))
+        subprocess.run(["python3",parent_dir+"/train/train.py"])
+    else:
+        print("bag2labels.py failed")
